pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Maksim <knyazev.maksim2852@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
python-dotenv = "^1.0.1"
geopy = "^2.4.1"
alembic = "^1.13.3"
aiogram = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
